[fix] this should be a float

+++ js/api.go
@@ -123,6 +123,6 @@ func (a JSAPI) Taint() {
 	a.vu.Taint = true
 }
 
-func (a JSAPI) ElapsedMs() int64 {
-	return int64(time.Since(a.vu.started) * time.Millisecond)
+func (a JSAPI) ElapsedMs() float64 {
+	return float64(time.Since(a.vu.started)) * float64(time.Millisecond)
 }
